+++ src/lib/components/layout/Sidebar.svelte
@@ -1,11 +1,14 @@
 <script lang="ts">
-  import { setImportedNodes } from '../../states/store.svelte'
+  import { getImportedNodes, setImportedNodes } from '../../states/store.svelte'
   import { useDnD } from '../DnDProvider.svelte'
   import defaultNodes from '../../data/defaultNodes.json'
-  import type { ImportedNodes } from '../../types/nodeTypes'
+  import { nodeColors, type NodeData } from '../../types/nodeTypes'
 
-  let importedNodes: ImportedNodes | {} = $state(defaultNodes)
-  setImportedNodes(defaultNodes)
+  if (defaultNodes) {
+    // TODO: add stantilization checks (i.e. empty object) and move into separate function
+    setImportedNodes(defaultNodes)
+  }
+  const availableNodesByType = $derived(getImportedNodes())
 
   const type = useDnD()
 
@@ -17,49 +20,32 @@
     event.dataTransfer.effectAllowed = 'move'
   }
 
-  const onFileChange = async (e) => {
-    const file = e.target.files[0]
-    if (file == null) {
-      return
-    }
-    importedNodes = await readJsonFile(file)
-    setImportedNodes(importedNodes)
+  const returnNodeType = (node) => {
+    return 'method_name' in node ? node.method_name : node.type
   }
-
-  const readJsonFile = (file) => {
-    const reader = new FileReader()
-    return new Promise((resolve, reject) => {
-      reader.onload = () => resolve(JSON.parse(reader.result as string))
-      reader.onerror = reject
-      reader.readAsText(file)
-    })
+  const returnNodeColor = (nodeTypeName) => {
+    return nodeColors[nodeTypeName]
   }
-
-  const returnNodeType = (node) =>
-    'method_name' in node ? node.method_name : node.type
 </script>
 
 <aside>
   <div class="nodes-container">
-    <div class="label">
-      Upload an import file to load the nodes to be drag in the canvas below.
-    </div>
-    <input type="file" onchange={onFileChange} accept=".json" />
-  </div>
-  <div class="nodes-container">
-    {#if importedNodes}
-      {#each Object.keys(importedNodes) as nodeId (nodeId)}
-        <div
-          role="option"
-          aria-selected="false"
-          tabindex="0"
-          class="node"
-          ondragstart={(event) =>
-            onDragStart(event, returnNodeType(importedNodes[nodeId]))}
-          draggable={true}
-        >
-          {returnNodeType(importedNodes[nodeId])}
-        </div>
+    {#if availableNodesByType}
+      <!-- TODO: move into separate function -->
+      {#each Object.entries(availableNodesByType) as [nodeTypeName, arrNodesByType] (nodeTypeName)}
+        {#each arrNodesByType as Array<NodeData> as node (node)}
+          <div
+            style="--borderColor: {returnNodeColor(nodeTypeName)}"
+            role="option"
+            aria-selected="false"
+            tabindex="0"
+            class="node"
+            ondragstart={(event) => onDragStart(event, returnNodeType(node))}
+            draggable={true}
+          >
+            {returnNodeType(node)}
+          </div>
+        {/each}
       {/each}
     {/if}
   </div>
@@ -67,18 +53,9 @@
 
 <style>
   aside {
-    width: 100%;
-    background: #efefef;
-    padding: 1rem;
-    font-size: 12px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .label {
-    font-size: 0.9rem;
+    height: 100vh;
+    background: var(--background-color-secondary);
+    font-size: 1rem;
   }
 
   .nodes-container {
@@ -86,15 +63,21 @@
     flex-wrap: wrap;
     align-items: center;
     justify-content: center;
+    overflow-y: auto; /* Vertical scrollbar only when overflowing */
     gap: 1rem;
-    margin-top: 1rem;
+    padding: 2rem 1rem;
   }
 
   .node {
-    border: 1px solid #111;
     padding: 0.5rem 1rem;
-    font-weight: 700;
     border-radius: 5px;
     cursor: grab;
+    border: 2px solid var(--borderColor, gray);
+    color: var(--text-color);
+    background-color: var(--background-secondary-color);
+  }
+
+  .node:hover {
+    border-color: var(--border-color-hover);
   }
 </style>